Replace validate_prefs() with check_prefs()
Don't try to parse a prefs.js file to check if it is valid. Instead
take the prefs.js file that is output by the browser after launch
and check if the prefs defined in the external prefs file were
merged.

This will have some false positives if the external prefs file
does not stick to the format output by the browser. Updating the
external prefs.js file to avoid false positives is simple and
worth it.